Created the solution with the Api Project and Database Project.
Created the repository. I will work in main since noone else will be working on it. Set it to public so I can share it.
Created 2 folders. One for the .Net Backend solution, one for the Angular solution. This is not ideal, each would have to have it's own.
I created the first project for the .Net solution, its a library for storing the DB Models and Context in EF.
Added a DBEntity class thar would be used for the audit data, in this case just CreateDate and UpdateDate.
Created the Customer class.
Created the DBContext.
Added the API solution for the first project. Deleted the example controller and class.
Added the Database project as reference. Added EF Designer so I can create migrations. Added the connection String to appsettings. Added the DBContext to the DI Services pointing to the correct project.
LEARN: Couldn't create the migration first. The problem was that the API project was not the starting solution.
I created a customer as Seed data just to check. I also made sure in the seeding that the Database is created and to apply any pending migration. This would be not the way to go in a real app.