Implemented the API with an Application Layer. Implemented some Unit Tests.
Wanted to be able to check the assembly version from an API call, and it to auto-increment. I had to add two options to the API Project definition so it doesn't generate the information by itself. Then I used the [assembly: AssemblyVersion("1.0.*")] attribute in Program.cs.
I'm going to use an application layer and communicate with it using Mediator pattern CQRS. Should I create a DDD model? or is an overkill for this? I decided to go with a simplified CleanArchitecture, where Database is an Infraestructure Layer, Application will be the Business Logic Layer, and API is the UI Layer.
The Application Layer is the only one that touches both UI and Database. The communication between UI and Application will be done using the Mediator pattern via MediatR Library and a simple CQRS.
I will not use a complete DDD Domain, this would be considered an Anemic Domain Model but for a solution with just one entity this is allready an overkill.
Since the Logic Layer is allready using the Database project as an Anemic Domain, I'm also skiping the repository pattern.
Created the Application Project for that layer.
Created the controller for Customers.
Started with GetAll. I decided to forgoe a DTO specific for the API since the Responses are very straight foward. But I decided I should implement some paging, sorting and filtering. I created a PagedList object to return to the API with additional data.
Implemented the QueryRequest and Handler. Implemented Mapperly to map between Customer and CustomerResponse.
Implemented a Unit test to be sure the HasNext and HasPrevious methods of the PageList are working. I'm using an InMemoryDB and Boggle to generate some fake data.
Implmented POST, PUT, PATCH and DELETE. I would have skipped Patch but I implemented Put as Patch and since It was allready there I just renamed it.
I reused CustemrResponse for the rest of the calls since is enough for now.
I'm skipping Unit test since it's veen a while now.
I haven't committed in a while.