Implemented the Customer List
Implemented the models for CustomerResponse and PageList, the CustomerService with the GetAll function, and the view in plan html. But I don't like it. I decided I needed a template.
I found a template form PrimeNG and was actually easy to implement. It has a lot of pages but the libraries are clean and don't any clutter like jquery or bootstrap.
Here I made my list using the PrimeNG DataTable component. Had to add *ngIf for LastUpdated because it crashed when it's null.