Set controllers in an external project.
Next we are taking the controllers to an external dependency so they don't have access to the DBcontext, not even by mistake.

I passed all the controller classes. Try to build.

There is an issue, that all the base class libraries from AspNetCore.App are not available in Nuget anymore. To add them you have o add a framework reference, but to do this you should do it directly in the csproj file. Adding this:
<ItemGroup>
    <FrameworkReference Include="Microsoft.AspNetCore.App" />
  </ItemGroup>

Now, to add the controllers from another library, you have to change Program.cs and register them from an external assembly.
builder.Services.AddControllers().AddApplicationPart(typeof(VersionController).Assembly);