Implemented CustomerResponse contract in the rest of the API calls.
While implementing the same CustomerResponse contract in the rest of the API calls, one problem appeared. When the result is null, the mapper is not able to work since it is expecting an object. I defined both, the response and the parameter for that method so it can handle nulls by returning another null value.

With this all the API calls work again, but now they respond with an API contract instead of the Results from the Application Layer.