Implmeneted Post and Put Calls with it's own DTO. And Added Unit Tests for both actions.
After a lot of testing, it looks that mixing DataAnnotations with Record types as input for Controllers might have some issues.
It started when I tried to change the POST and PUT input parameters for a DTO using the Record Type.
Firts I created the record type, to add DataAnnotations to the constructor declared properties you have to add them while declaring the property.
```
public record CreateCustomerRequest(
            [property: Required, StringLength(200, MinimumLength = 2)] string firstName,
            [property: Required, EmailAddress] string email,
            [property: StringLength(200)] string? lastName = "",
            [property: StringLength(100)] string? phone = "",
            [property: StringLength(10000)] string? description = ""
        )
    {}
```
I decided then to test the controllers. But I kept having this error:
```
System.InvalidOperationException: Record type 'NetBackend_Api_Controllers.Contracts.CreateCustomerRequest' has validation metadata defined on property 'description' that will be ignored. 'description' is a parameter in the record primary constructor and validation metadata must be associated with the constructor parameter.
It didn't matter if I sent that property with data or not.
```
In the unit test that check for the controller to send a Data Annotation error worked fine. Then I removed the word property:, and the Unit test stopped working and the controller worked.
Notice the property tag in the annotation, its necessary for using Validators. But using it, breakes the controller.
I tried manually creating the Record with it's constructor and setting the Annotation Attributes on the properties, but this is no good, the annotation have to be in the ctor.
This page also describes the same behavior:
https://github.com/dotnet/runtime/issues/49011

Basically it says to change it for classes. They don't plan to add the validation attributes. I changed CreateCustomerRequest to that then.